Allow passing arbitrary options to the http client
This change is backwards compatible. Constructing a Mailchimp object was
previously done this way:

    new Mailchimp($key, $user, $timeout);

This commit introduces a more flexible approach:

    new Mailchimp($key, $user, $http_options);

For backwards compatibility, $http_options can be an integer indication
the desired request timeout.

If $http_options is an array without a 'timeout' element, the default
timeout of 10 seconds is added to the http options. Otherwise the
options are passed verbatim to the guzzle Client constructor.

Typical options a user of this library might want to pass to the http
client are debug settings, proxies, connection timeouts or ssl options.

Fixes #32.